move graphql schema out of main
